requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0